Add module for uploading to Sketchfab
The module uses OAuth to log in to Sketchfab and the REST API to
upload files. Config variables were added for this. When uploading a
file, a separate task is created for each site it is uploaded to. The
log in button for Sketchfab is shown when the user has logged in to
Commons, as the latter is used by 3D2commons to keep track of users.